requirements.txt
paho-mqtt
requests
structlog